Changed how data is stored
Instead of loading datasets from a JSON file into a JSON object, datasets are now parsed into the dataSet object. Meaning all the dataSet specific methods should do something now. To add a consequence of this means instead of saving the data sets as a JSONObject of JSONObjects, now it's stored as an JSONArray of JSONObjects.